+++ src/components/document-viewer.tsx
@@ -12,19 +12,7 @@ import type { AdviceAPIType } from "@/types";
 import OpenInternalLink from "./internal-link";
 import OpenExternalLink from "./external-link";
 import type { UsersTableType } from "@/server/db/schema";
-import { Download } from "lucide-react";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog";
-import { toast } from "sonner";
+
 
 const queryClient = new QueryClient();
 
@@ -95,40 +83,9 @@ function SkeletonLoader() {
   );
 }
 
-const fetchUserData = async (): Promise<UsersTableType> => {
-  const response = await fetch(`/api/getUserData`);
-  if (!response.ok) {
-    throw new Error("Network error");
-  }
-  const data = (await response.json()) as UsersTableType;
-  return data;
-};
 
-async function deleteDocument(documentId: number) {
-  const userData = await getUserData();
-  if (!userData) {
-    throw new Error("User data not found");
-  }
-  const body = {
-    documents: [documentId],
-    method: "delete",
-  };
-  const response = await fetch(
-    `${userData.paperlessURL}/api/documents/bulk_edit/ `,
-    {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Token ${userData.paperlessToken}`,
-      },
-      body: JSON.stringify(body),
-    },
-  );
-  return response;
-}
-
-function DocumentViewer(props: { id: number }) {
-  const router = useRouter();
+function DocumentPreview(props: { id: number }) {
+    const router = useRouter();
 
   const { data: userData, isLoading: isUserDataLoading } = useQuery({
     queryKey: ["userData"],
@@ -156,8 +113,7 @@ function DocumentViewer(props: { id: number }) {
       </div>
     );
   }
-
-  return (
+    return (
     <div className="flex h-full w-full min-w-0 justify-center">
       <div className="flex h-4/5 flex-col rounded-xl bg-slate-600/50 md:w-1/2">
         <div className="m-4 flex flex-grow flex-col justify-center gap-8 md:m-8 md:flex-row md:gap-16">
@@ -167,84 +123,16 @@ function DocumentViewer(props: { id: number }) {
               type="application/pdf"
               width="100%"
               height="100%"
-            >
-              <p>
+            >              <p>
                 Your web browser doesn&apos;t have a PDF plugin. Instead you can
                 <OpenInternalLink href={pdfUrl}>
                   click here to download the PDF file.
                 </OpenInternalLink>
               </p>
             </object>
-          </div>
-          <div className="flex flex-col gap-8">
-            <Button
-              onClick={(e) => {
-                e.preventDefault();
-                router.back();
-              }}
-            >
-              Back
-            </Button>
-            <a
-              href={pdfUrl}
-              download={pdfUrl}
-              className={`${buttonVariants({ variant: "default" })}`}
-            >
-              Download
-              <Download />
-            </a>
-            <div id="dialog-container" />
-            <OpenExternalLink
-              className={buttonVariants({ variant: "default" })}
-              href={`${userData.paperlessURL}/documents/${props.id}/details/`}
-            >
-              Open
-            </OpenExternalLink>
-            <AlertDialog>
-              <AlertDialogTrigger>
-                <Button className="w-24" variant="destructive">
-                  Delete
-                </Button>
-              </AlertDialogTrigger>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    This action cannot be undone. This will permanently delete
-                    the recording.
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <AlertDialogCancel>Cancel</AlertDialogCancel>
-                  <AlertDialogAction
-                    onClick={async () => {
-                      const response = await deleteDocument(props.id);
-                      if (response.ok) {
-                        toast("Pdf deleted", {
-                          description: "The recording has been deleted.",
-                        });
-                      } else {
-                        toast("Error deleting pdf", {
-                          description:
-                            "An error occurred while deleting the pdf.",
-                        });
-                      }
-                      router.back();
-                    }}
-                  >
-                    Continue
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
+            )}
 
-export default function Page(props: { id: number }) {
+export default function DocumentViewer(props: { id: number }) {
   return (
     <QueryClientProvider client={queryClient}>
       <DocumentViewer id={props.id} />
